better image conversion
git-svn-id: svn://svn.code.sf.net/p/rtf2latex2e/code@247 ec8729b6-1b83-4082-b6c8-10560a4245e1

+++ src/rtf2latex2e.h
@@ -70,8 +70,8 @@ typedef struct
     int32_t height;
     int32_t goalWidth;
     int32_t goalHeight;
-    int32_t scaleX;
-    int32_t scaleY;
+    float scaleX;
+    float scaleY;
     int     llx;
     int     lly;
     int     urx;
